Das Interface GameObject wurde durch eine abstrakte Klasse ersetzt. Umbenennung von Objekten für bessere Lesbarkeit.
