+++ src/main/java/de/mth/game/gameobject/Player.java
@@ -11,7 +11,7 @@ import de.mth.game.common.Input;
 import de.mth.game.texture.Animation;
 import de.mth.game.texture.TextureLoader;
 
-public class Player extends GameObjectBase {
+public class Player extends GameObject {
 
 	private Animation walkUp;
 	private Animation walkDown;
